requirements.txt
# asgiref @ file:///home/runner/.cache/pypoetry/artifacts/44/47/8b/eeb8b3236803e39b50f24aec1dd7dc622a47aedd27d0c37ba347271f9b/asgiref-3.5.1-py3-none-any.whl
backports.entry-points-selectable==1.1.1
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cleo==0.8.1
click==8.1.3
clikit==0.6.2
colorama==0.4.6
crashtest==0.3.1
crayons==0.4.0
cryptography==35.0.0
debugpy==1.2.1
distlib==0.3.3
Django==3.2.13
docopt==0.6.2
filelock==3.3.2
future==0.18.2
html5lib==1.1
idna==3.3
importlib-metadata==3.10.1
jedi==0.15.2
jeepney==0.7.1
keyring==21.8.0
lockfile==0.12.2
msgpack==1.0.2
numpy==1.22.2
opencv-python==4.6.0.66
packaging==20.9
parso==0.5.2
pastel==0.2.1
pexpect==4.8.0
piexif==1.1.3
Pillow==9.3.0
pipreqs-amasad==0.4.10
pkginfo==1.7.1
platformdirs==2.4.0
pluggy==0.13.1
poetry==1.1.11
poetry-core==1.0.7
ptyprocess==0.7.0
pycparser==2.21
pyflakes==2.1.1
pylev==1.4.0
pyparsing==3.0.5
python-jsonrpc-server==0.3.2
python-language-server==0.31.10
# pytz @ file:///home/runner/.cache/pypoetry/artifacts/2f/64/84/b6333ff5a89076f2f48899fb6dc7e562400c812c962713b90b44ac5160/pytz-2022.1-py2.py3-none-any.whl
requests==2.26.0
requests-toolbelt==0.9.1
rope==0.18.0
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
# sqlparse @ file:///home/runner/.cache/pypoetry/artifacts/5c/57/47/582e6c575570333d1c84c495489bc30ed2c7e0fccaabeaa29650cfc543/sqlparse-0.4.2-py3-none-any.whl
stegano==0.10.2
stego-lsb==1.3.5
toml==0.10.2
tomlkit==0.7.2
typing-extensions==3.7.4.3
ujson==1.35
urllib3==1.26.7
virtualenv==20.10.0
webencodings==0.5.1
yapf==0.31.0
yarg==0.1.9
zipp==3.4.1
